added index.ts files for directories missing them
