+++ src-ui/index.ts
@@ -133,7 +133,7 @@ window.addEventListener(NoteEvents.Selected, async (event) => {
     await renderNoteEditor({ isLoading: false, note })
 
     // based on URL params, render dialogs
-    // note: this could potentially be moved to a `DialogEvents.Open` with which dialog to render passed in
+    // note: this could potentially be moved to a `DialogEvents.Open` with which dialog to render passed in --- GOOD IDEA!
     switch (dialog) {
       case 'details':
         note && renderNoteDetailsDialog(note)
